requirements.txt
aiogram
fastapi
uvicorn
python-dotenv
motor
